ErrorMonitor implementation more like NetworkMonitor
Removed all but the message.stateIn from AppState

Added interface and implementation, dependency injection
